feat: Implement Meal Book Update and Delete Functionality
This commit introduces the ability to update and delete meal books directly from the Meal Screen, enhancing user control over their meal management.

*   **Meal Book Update:**
    *   Adds `EditMealDialog` for updating meal book details, including name, default price, and description.
    *   Implements `UpdateMealBook` event in `MealScreenEvents` to handle update requests.
    *   Introduces a long-press action on `MealItem` to trigger the `EditMealDialog`.
    *   Uses `UpdateMealBook` use case to update meal book details in the database.
    *   Shows a toast message upon successful update or if the update fails.
    *   Updates the `Add` to `Update` in `EditMealBookDialog`.
    * Updates `AddMealBookState` with `mealBookId`
    * Update `MealDao` to return the number of row updated with `updateMealBook`.
    * Update `UpdateMealBook` to return the number of row updated.
    * Update the  `mapFromEntity` and `mapToEntity` from `AddMealBookMapper`
*   **Meal Book Deletion:**
    *   Adds a delete option within `EditMealDialog` to remove meal books.
    *   Implements `DeleteMealBook` event in `MealScreenEvents` to handle deletion requests.
    *   Uses `DeleteMealBook` use case to remove meal book from the database.
    *   Shows a confirmation dialog before deleting a meal book.
    *   Displays a toast message upon successful deletion.
    * Adds `Delete` Icon in `EditMealBookDialog`
    * Adds `Delete` Icon in `EditMealDialog`
*   **UI Enhancements:**
    *   Replaces `EditMealBookDialog` with `EditMealDialog` for editing.
    * Rename `EditMealBookDialog` to `EditMealDialog`
    * Rename `editmealbook` package to `edit` package
    * Adds `combinedClickable` for click and long click in `MealItem`
    * Update the state name in `EditMealDialog`.
    * Updates the `AlertDialog` for showing the `Delete` button for delete the meal book.
    * Updates `AlertDialog` for showing the `Edit` Icon for editing the meal book.
    * Adds `Icons.TwoTone.Delete` in `AlertDialog`.
*   **Code:**
    *   Adds `EditMealDialog` composable.
    *   Adds `DeleteMealBook` and `UpdateMealBook` events in `MealScreenEvents`.
    *   Updates `MealViewModel` to handle `DeleteMealBook` and `UpdateMealBook` events.
    * Adds a `isEditDialogVisible` for the edit dialog.
    * Adds `yetEditModel` for handling the edit model.
    * Adds `onlLongClick` to `MealItem` for the long press action.
    * Adds the logic for showing the dialog in the `MealScreen`.
    * Adds the `EditMealDialog` to the `MealScreen`.
    * Adds the `onDeleteItem` in the `EditMealDialog` for the delete action.
    * Adds the `confirmButton` in the `EditMealDialog` for the update action.
    * Update the `state` value.
    * Update the `price` value.
    * Adds `isDeleteDialogVisible` for showing the delete dialog.
    * Adds the `Delete` icon in the `EditMealDialog`.
    * Adds the `Edit` icon in the `EditMealDialog`.
    * Adds a icon for the delete dialog.

+++ shared/src/commonMain/kotlin/com/atech/expensesync/usecases/MealBookUseCases.kt
@@ -35,7 +35,7 @@ data class GetMealBookEntries(private val dao: MealDao) {
 }
 
 data class UpdateMealBook(private val dao: MealDao) {
-    suspend operator fun invoke(mealBook: MealBook) = dao.updateMealBook(mealBook)
+    suspend operator fun invoke(mealBook: MealBook) = dao.updateMealBook(mealBook).toLong()
 }
 
 data class UpdateMealBookEntry(private val dao: MealDao) {
